refactor(ast): extract schemaSignature/isSchemaEqual into signature.ts
Move the shape-only signature (hash-consing) and equality check out of
dedupe.ts into a focused signature.ts; dedupe.ts imports signatureOf
from it. Split the signature tests into signature.test.ts.

Also broaden adapter.test.ts to inline snapshots for the stream schema
list, operation ids, stream meta, getImports result, and the deduped
schema sets.
