+++ packages/ast/src/index.ts
@@ -1,5 +1,5 @@
 export { httpMethods, isScalarPrimitive, mediaTypes, nodeKinds, schemaTypes } from './constants.ts'
-export { applyDedupe, buildDedupePlan, isSchemaEqual, schemaSignature } from './dedupe.ts'
+export { applyDedupe, buildDedupePlan } from './dedupe.ts'
 export { defineSchemaDialect } from './dialect.ts'
 export { dispatch } from './dispatch.ts'
 export {
@@ -35,6 +35,7 @@ export { isHttpOperationNode, isInputNode, isOperationNode, isOutputNode, isSche
 export { createPrinterFactory, definePrinter } from './printer.ts'
 export { extractRefName } from './refs.ts'
 export { childName, collectImports, enumPropName, findDiscriminator } from './resolvers.ts'
+export { isSchemaEqual, schemaSignature } from './signature.ts'
 export { mergeAdjacentObjects, mergeAdjacentObjectsLazy, setDiscriminatorEnum, setEnumName, simplifyUnion } from './transformers.ts'
 export type * from './types.ts'
 export {
